Update iframe.html
In the case that an on premise JAMF Pro server exists, the URL within the iframe.html on line 111 contains a reference to ".jamfcloud.com." This will result in a blank page when clicking the "View User in Jamf Pro" link within Zendesk. Removing the reference to .jamfcloud.com is necessary for on premise users of Jamf Pro. I would recommend two releases of the Zendesk application, one for hosted/on premise users and one for Jamf Cloud users, each containing the appropriate HTML code for each.